Update magazine folder name
change "subscription" URL to "magazine". need to move magazine to "issue"
