+++ html/issue/.htaccess
@@ -2,7 +2,7 @@
 Options -Multiviews
 
 RewriteEngine on
-RewriteBase /magazine/
+RewriteBase /issue/
 RewriteRule ^([0-9]{4}/[0-9]{2})$ $1/ [L,R=301]
 RewriteRule ^([0-9]{4}/[0-9]{2}/)$ index.php?f=$1index.html [L]
 RewriteRule ^([0-9]{4}/.*)$ index.php?f=$1 [L]
