+++ html/magazine/ajax/updatecard/index.php
