+++ inc/layout/header1.php
@@ -21,7 +21,7 @@ print'
                 <li class="nav-link"><a href="'.$protocol.$site.'/contact/">Stay In Touch</a></li>
                 <li class="nav-link"><a href="http://helmarblog.com/" target="_blank">Blog</a></li>
                 <li class="nav-link"><a href="http://stores.ebay.com/Helmar-Brewing-Art-and-History/" target="_blank">eBay Store</a></li>
-                <li class="nav-link"><a href="'.$protocol.$site.'/subscription/">Magazine</a></li>
+                <li class="nav-link"><a href="'.$protocol.$site.'/magazine/">Magazine</a></li>
 ';
 if(isset($user)){
     if( $user->login() == 1 || $user->login() == 2 ){
